revert timestamp to double

+++ src/ray/gcs/pb_util.h
@@ -33,7 +33,7 @@ inline std::shared_ptr<ray::rpc::JobTableData> CreateJobTableData(
 /// Helper function to produce error table data.
 inline std::shared_ptr<ray::rpc::ErrorTableData> CreateErrorTableData(
     const JobID &job_id, const std::string &error_type, const std::string &error_msg,
-    int64_t timestamp) {
+    double timestamp) {
   auto error_info_ptr = std::make_shared<ray::rpc::ErrorTableData>();
   error_info_ptr->set_job_id(job_id.Binary());
   error_info_ptr->set_type(error_type);
